[ADD]Error message for proyeccines

+++ ctm_sentencias/models/proyecciones.py
@@ -37,6 +37,9 @@ class Proyecciones(models.Model):
 
     def calcular_proyeccion(self):
         for record in self:
+            if record.sentencia_id.fecha_liquidar <= record.sentencia_id.fecha_liquidar_optimista:
+                raise ValidationError('La fecha de liquidación no puede ser menor o igual a la fecha de liquidación optimista')
+
             record.liquidacion_inicial_ids.unlink()
             record.generar_liquidacion_inicial()
             # Resultados
